plantlisting-backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
bcrypt==3.2.0
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.2
defusedxml==0.6.0
distlib==0.3.1
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.1
django-address==0.2.5
django-braces==1.14.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
django-location-field==2.1.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
docker==4.4.1
dockerpty==0.4.1
docopt==0.6.2
facebook-sdk==3.1.0
filelock==3.0.12
gunicorn==20.0.4
heroku==0.1.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.1.1
importlib-resources==3.3.0
incremental==17.5.0
jsonschema==3.2.0
msgpack==0.6.2
numpy==1.19.4
oauthlib==3.1.0
paramiko==2.7.2
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.1.0
PyNaCl==1.4.0
pyOpenSSL==20.0.0
pyotp==2.4.1
pyrsistent==0.17.3
python-dateutil==1.5
python-decouple==3.3
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
scipy==1.5.4
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
texttable==1.6.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
websocket-client==0.57.0
whitenoise==5.2.0
zipp==3.4.0
zope.interface==5.2.0
